Script to count appearances of Harry and Tom Riddle :D
This script must be run with: awk -f harry_voldemort.awk Harry\ Potter\ 1\ -\ Harry\ Potter\ and\ the\ Sorcerer\'s\ Stone\ -\ J.\ K.\ Rowling\ \&\ Mary\ Grandpre.txt 

Due to a bad choice on the file name (spaces on it) we can't create a bash loop to run with all the files. But we could run it with perl or creating another script to handle this problem.